+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## ?.?.?
+## 1.4.0
 
 * New commands `teams add-member` and `teams remove-member` added, allowing for
   management of team members.